refactor(agent-task): split scheduler plan types
